Moved JSONEncoder to separate file
